requirements.txt
fastapi
uvicorn
sse-starlette
beautifulsoup4
google-generativeai
requests
fastmcp
flask

#pip install -r requirements.txt